Move account fragment to fragment package
